Refactor a model analysis part out of ExtendedRequest
ExtendedRequest contained quite a bit of logic for scanning a model
class for query fields and joins.  In this commit, it is separated out
into a `DBEntityAnalsisBuilder` (to do the scanning) and a
`DBEntityAnalysis` (which is the result).

The `DBEntityAnalsysis` has been made immutable for simplicity.  In
theory, it also enables us to reuse the `DBEntityAnalysis` result
between multiple entities but in practise this is not very useful
right now (as `ExtendedRequest` are currently single use items).
