+++ src/main/java/org/dcsa/core/extendedrequest/JoinDescriptor.java
@@ -1,33 +1,20 @@
 package org.dcsa.core.extendedrequest;
 
+import org.springframework.data.relational.core.sql.Join;
+
 public interface JoinDescriptor {
 
-    org.springframework.data.relational.core.sql.Join.JoinType getJoinType();
+    Join.JoinType getJoinType();
     String getTableName();
     String getJoinAlias();
     String getJoinCondition();
     String getDependentAlias();
-    boolean isInUse();
-    void setInUse(boolean inUse);
 
-    default void apply(Join join) {
+    default void apply(Joiner joiner) {
         String tableName = getTableName();
         String joinAlias = getJoinAlias();
         String joinCondition = getJoinCondition();
         String joinLine = tableName + " AS " + joinAlias + " " + joinCondition;
-        org.springframework.data.relational.core.sql.Join.JoinType joinType = getJoinType();
-        switch (joinType) {
-            case JOIN:
-                join.doInner(joinLine);
-                break;
-            case LEFT_OUTER_JOIN:
-                join.doLeft(joinLine);
-                break;
-            case RIGHT_OUTER_JOIN:
-                join.doRight(joinLine);
-                break;
-            default:
-                throw new IllegalArgumentException("Unsupported joinType: " + joinType.name());
-        }
+        joiner.addJoin(this, joinLine);
     }
 }
