+++ src/test/java/org/dcsa/core/models/CitySpecificExtendedRequest.java
@@ -1,6 +1,8 @@
 package org.dcsa.core.models;
 
 import org.dcsa.core.extendedrequest.*;
+import org.dcsa.core.query.DBEntityAnalysis;
+import org.dcsa.core.util.ReflectUtility;
 import org.springframework.data.relational.core.sql.Join;
 
 public class CitySpecificExtendedRequest extends ExtendedRequest<City> {
@@ -8,25 +10,18 @@ public class CitySpecificExtendedRequest extends ExtendedRequest<City> {
         super(extendedParameters, City.class);
     }
 
-    public Class<?> getPrimaryModelClass() {
-        return this.getModelClass();
-    }
-
-    protected void findAllTablesAndBuildJoins() {
-        super.findAllTablesAndBuildJoins();
+    protected DBEntityAnalysis.DBEntityAnalysisBuilder<City> prepareDBEntityAnalysis() {
         JoinDescriptor joinDescriptor = SimpleJoinDescriptor.of(
                 Join.JoinType.JOIN,
-                getTableName(County.class),
+                ReflectUtility.getTableName(County.class),
                 "c",
-                "ON c.id = " + getTableName(getModelClass()) + ".country_id",
+                "ON c.id = " + ReflectUtility.getTableName(getModelClass()) + ".country_id",
                 null
         );
-        registerJoinDescriptor(joinDescriptor);
-    }
-
-    protected void loadFieldsFromSubclass() {
-        registerQueryField(QueryFields.nonSelectableQueryField(
-                "c", "country_name", "cn", String.class
-        ));
+        return super.prepareDBEntityAnalysis()
+                .registerJoinDescriptor(joinDescriptor)
+                .registerQueryField(QueryFields.nonSelectableQueryField(
+                        "c", "country_name", "cn", String.class
+                ));
     }
 }
